projects/opendiscourse-mcp/servers/mcp_govdocs/requirements.txt
fastapi==0.112.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
psycopg[binary]==3.2.1
httpx==0.27.0
beautifulsoup4==4.12.3
lxml==5.3.0
readability-lxml==0.8.1
pdfminer.six==20240706
redis==5.0.7
minio==7.2.7

psycopg_pool==3.2.1
tenacity==8.5.0

PyYAML==6.0.2
